+++ CHANGELOG.md
@@ -7,9 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased] 
 
-### Fixed
-
-- When adding to existing graph the source graph IRI is added to the label instead of replacing the source graph IRI
+- When adding to existing graph the source graph IRI is added to the label
 
 ### Added